feat: Refactor engine classes to inherit from SearchEngineBase and streamline logic
